+++ build.zig
@@ -239,7 +239,7 @@ const module_list = [_]Module{
     .{ .name = "json5", .libs = &.{"format"}, .test_deps = &.{"testkit"} },
     .{ .name = "yaml", .libs = &.{"format"} },
     .{ .name = "jinja", .libs = &.{"format"}, .test_deps = &.{"testkit"} },
-    .{ .name = "cbor", .libs = &.{"format"} },
+    .{ .name = "cbor", .libs = &.{"format"}, .test_deps = &.{"testkit"} },
     .{ .name = "protobuf", .libs = &.{ "format", "web" } },
     .{ .name = "grpc", .libs = &.{"web"}, .deps = &.{ "http", "protobuf" }, .test_deps = &.{"testkit"} },
     .{ .name = "webauthn", .libs = &.{"crypto"}, .deps = &.{ "cbor", "rsa", "p256", "x509" }, .test_deps = &.{"testkit"} },